Rename src/node.js -> validator-node.js
